requirements.txt
chardet==5.2.0
numpy==2.2.5
pandas==2.2.3
pymysql==1.1.1
python-dotenv==1.1.0
Requests==2.32.3
selenium==4.32.0
SQLAlchemy==2.0.40
tqdm==4.67.1
